move data sync to DailyCheckService
also refactor of the NetworkFragment, since a service can't have a Fragment
