+++ app/src/main/java/de/beusterse/abfalllro/service/NetworkClient.java
@@ -1,13 +1,8 @@
 package de.beusterse.abfalllro.service;
 
-import android.app.Fragment;
-import android.app.FragmentManager;
-import android.content.Context;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.AsyncTask;
-import android.os.Bundle;
-import android.support.annotation.Nullable;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -27,59 +22,14 @@ import de.beusterse.abfalllro.interfaces.DownloadCallback;
  * Created by Google Samples
  * https://github.com/googlesamples/android-NetworkConnect/blob/master/Application/src/main/java/com/example/android/networkconnect/NetworkFragment.java
  */
-public class NetworkFragment extends Fragment {
-    public static final String TAG = "NetworkFragment";
-
-    private static final String URL_KEY = "UrlKey";
-
+public class NetworkClient {
     private DownloadCallback mCallback;
     private DownloadTask mDownloadTask;
     private String mUrlString;
 
-    /**
-     * Static initializer for NetworkFragment that sets the URL of the host it will be downloading
-     * from.
-     */
-    public static NetworkFragment getInstance(FragmentManager fragmentManager, String url) {
-        NetworkFragment networkFragment = (NetworkFragment) fragmentManager
-                .findFragmentByTag(NetworkFragment.TAG);
-        if (networkFragment == null) {
-            networkFragment = new NetworkFragment();
-            Bundle args = new Bundle();
-            args.putString(URL_KEY, url);
-            networkFragment.setArguments(args);
-            fragmentManager.beginTransaction().add(networkFragment, TAG).commit();
-        }
-        return networkFragment;
-    }
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setRetainInstance(true);
-        mUrlString = getArguments().getString(URL_KEY);
-        mCallback.ready();
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        // Host Activity will handle callbacks from task.
-        mCallback = (DownloadCallback) context;
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        // Clear reference to host Activity to avoid memory leak.
-        mCallback = null;
-    }
-
-    @Override
-    public void onDestroy() {
-        // Cancel task when Fragment is destroyed.
-        cancelDownload();
-        super.onDestroy();
+    public NetworkClient(DownloadCallback callback, String url) {
+        mCallback = callback;
+        mUrlString = url;
     }
 
     /**
